+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/notification/common/exception/NotificationException.java
@@ -1,8 +1,7 @@
 package git_kkalnane.backend.starbucks.notification.common.exception;
 
 
-import git_kkalnane.backend.starbucks.auth.common.exception.AuthErrorCode;
-import git_kkalnane.backend.starbucks.global.error.core.BaseException;
+import git_kkalnane.backend.starbucks._global.error.core.BaseException;
 
 public class NotificationException extends BaseException {
